% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753.P1.R
\docType{package}
\name{MATH4753.P1}
\alias{MATH4753.P1}
\title{MATH4753.P1}
\description{
Subsets a given data set and produces plot with quadratic curve
}
\author{
Christian Hackelman \email{cphackelman@ou.edu}
}
